Commit made by the Bioconductor Git-SVN bridge. Consists of 8 commits.
Commit information:

Commit id: 5945c9202e735242649e1194f6e8315a5836f1ff
Commit message: changes to the external library c code interface
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 18:45:49 -0700
Author date: 2014-08-31 18:45:49 -0700

Commit id: 612f85fd6681f4cbab1d4f27e3669e35a68adf4b
Commit message: changes to size_t for median_log
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 18:32:29 -0700
Author date: 2014-08-31 18:32:29 -0700

Commit id: 0dccf18748a6c98dd1acbe8db34b77cca5fe53f5
Commit message: median changes for size_t
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 18:17:20 -0700
Author date: 2014-08-31 18:17:20 -0700

Commit id: e65b4f64dfa7940685ca21f28b56e4ee5300bb1e
Commit message: log_median changes for size_t
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 18:08:11 -0700
Author date: 2014-08-31 18:08:11 -0700

Commit id: b9984cf077c87cf39e02f1f488fe80a6ff43c91d
Commit message: fix mismatch between header and implementation
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 17:56:54 -0700
Author date: 2014-08-31 17:56:54 -0700

Commit id: 5a486521d978d49e95418286e1b348e66eb97395
Commit message: log_avg size_t changes
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 17:52:57 -0700
Author date: 2014-08-31 17:52:57 -0700

Commit id: 2675ccbfb8c1062c18426dd3e2cc2eba20eb9412
Commit message: biweight changes to update to size_t where appropriate
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 17:42:51 -0700
Author date: 2014-08-31 17:42:51 -0700

Commit id: 6a058f67ace9f88e78834e84932b71747c78abe6
Commit message: size_t rlated changes
Committed by: Ben Bolstad
Author Name: Ben Bolstad
Commit date: 2014-08-31 17:08:05 -0700
Author date: 2014-08-31 17:08:05 -0700



git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/preprocessCore@93724 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ inst/include/biweight.h
@@ -7,6 +7,8 @@
 #ifndef BIWEIGHT_H
 #define BIWEIGHT_H 1
 
+#include <stddef.h>
+
 /*! \brief log2 transform the data and then use a 1-step Tukey Biweight to summarize each column
  * 
  *  Given a data matrix of probe intensities compute average expression measure then log2 it and SE of this estimate
@@ -22,7 +24,7 @@
  *  
  */
 
-void tukeybiweight(double *data, int rows, int cols, double *results, double *resultsSE);
+void tukeybiweight(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
 
 /*! \brief Use a 1-step Tukey Biweight to summarize each column
  * 
@@ -39,7 +41,7 @@ void tukeybiweight(double *data, int rows, int cols, double *results, double *re
  *  
  */
 
-void tukeybiweight_no_log(double *data, int rows, int cols, double *results, double *resultsSE);
+void tukeybiweight_no_log(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
 
 /*! \brief Use a 1-step Tukey Biweight to summarize each column
  *
@@ -58,7 +60,7 @@ void tukeybiweight_no_log(double *data, int rows, int cols, double *results, dou
  *  
  */
 
-void TukeyBiweight(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes, double *resultsSE);
+void TukeyBiweight(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes, double *resultsSE);
 
 /*! \brief Use a 1-step Tukey Biweight to summarize each column
  *
@@ -76,7 +78,7 @@ void TukeyBiweight(double *data, int rows, int cols, int *cur_rows, double *resu
  *  
  */
 
-void TukeyBiweight_noSE(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes);
+void TukeyBiweight_noSE(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes);
 
 /*! \brief Use a 1-step Tukey Biweight to summarize each column
  *
@@ -93,7 +95,7 @@ void TukeyBiweight_noSE(double *data, int rows, int cols, int *cur_rows, double 
  *  
  */
 
-void TukeyBiweight_no_log_noSE(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes);
+void TukeyBiweight_no_log_noSE(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes);
 
 /*! \brief compute a 1-step Tukey Biweight
  *
@@ -106,6 +108,6 @@ void TukeyBiweight_no_log_noSE(double *data, int rows, int cols, int *cur_rows, 
  *
  */
 
-double Tukey_Biweight(double *x, int length);
+double Tukey_Biweight(double *x, size_t length);
 
 #endif
