+++ src/log_median.h
@@ -3,10 +3,10 @@
 
 
 
-void LogMedian(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes, double *resultsSE);
-void LogMedian_noSE(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes);
+void LogMedian(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes, double *resultsSE);
+void LogMedian_noSE(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes);
 
-void logmedian(double *data, int rows, int cols, double *results, double *resultsSE);
-void logmedian_no_copy(double *data, int rows, int cols, double *results, double *resultsSE);
+void logmedian(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
+void logmedian_no_copy(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
 
 #endif
